requirements.txt
aioredis==1.3.1
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
asyncio==3.4.3
attrs==22.1.0
autobahn==22.7.1
Automat==20.2.0
billiard==3.6.4.0
celery==5.2.7
certifi==2022.6.15
cffi==1.15.1
channels==3.0.5
channels-redis==3.4.1
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.5
constantly==15.1.0
cryptography==37.0.4
daphne==3.0.2
defusedxml==0.7.1
Deprecated==1.2.13
diff-match-patch==20200713
Django==4.0.8
django-celery-beat==2.3.0
django-celery-results==2.4.0
django-channels==0.7.0
django-import-export==2.8.0
django-timezone-field==5.0
et-xmlfile==1.1.0
hiredis==2.0.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
kombu==5.2.4
MarkupPy==1.14
msgpack==1.0.4
oauthlib==3.2.0
odfpy==1.4.1
openpyxl==3.0.10
packaging==21.3
Pillow==9.2.0
prompt-toolkit==3.0.31
psycopg2==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyOpenSSL==22.0.0
pyparsing==3.0.9
python-crontab==2.6.0
python-dateutil==2.8.2
pytz==2022.2.1
PyYAML==6.0
redis==4.3.4
requests==2.28.1
requests-oauthlib==1.3.1
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
tablib==3.2.1
Twisted==22.4.0
twisted-iocpsupport==1.0.2
txaio==22.2.1
typing_extensions==4.3.0
tzdata==2022.2
urllib3==1.26.12
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
xlrd==2.0.1
xlwt==1.3.0
zope.interface==5.4.0
